bot/modules/events: convert to TS
Convert bot/modules/eventsmodule to TypeScript.
